작업 내용 저장

+++ src/main/java/com/exam/gagi/model/Item.java
@@ -11,7 +11,7 @@ public class Item extends BaseEntity {
 	private String description; // 상품 설명
 	private String delivery; // 거래 지역/주소
 	private String isDirectDealAvailable; // 직거래 가능 여부 (Y/N)
-	private Long price; // 판매 가격
+	private BigDecimal price; // 판매 가격
 	private int amount; // 수량
 	private String bankAccountNumber; // 판매자 계좌번호
 	private String salesStatus; // 판매 상태
@@ -74,11 +74,11 @@ public class Item extends BaseEntity {
 		this.isDirectDealAvailable = isDirectDealAvailable;
 	}
 
-	public Long getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public void setPrice(Long price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
